pyproject.toml
[tool.poetry]
name = "db-connector"
version = "0.1.0"
description = ""
authors = ["guilherme <guilhermerezende770@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.12"
mysql-connector-python = "^8.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
